Refactor store into separate files
